fix: rename module to commands
