+++ packages/salesforcedx-vscode-lwc/src/commands/commandUtils.ts
@@ -5,7 +5,7 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-import { channelService } from '@salesforce/salesforcedx-utils-vscode/out/src/channels';
+import { channelService } from '@salesforce/salesforcedx-utils-vscode/out/src/commands';
 import * as vscode from 'vscode';
 import { nls } from '../messages';
