+++ packages/salesforcedx-vscode-lwc/test/vscode-integration/testSupport/testRunner/taskService.test.ts
@@ -9,7 +9,7 @@ import { assert, SinonStub, stub } from 'sinon';
 import * as vscode from 'vscode';
 import { taskService } from '../../../../src/testSupport/testRunner/taskService';
 import { nls } from '../../../../src/messages';
-import { ChannelService } from '@salesforce/salesforcedx-utils-vscode/out/src/channels';
+import { ChannelService } from '@salesforce/salesforcedx-utils-vscode/out/src/commands';
 
 describe('Task Service Unit Tests', () => {
   let executeTaskStub: SinonStub<
